Move OntoumlFactory class into its own file
