+++ __tests__/libs/ontouml2gufo/enumeration.test.ts
@@ -1,4 +1,5 @@
-import { OntoumlFactory, generateGufo } from './helpers';
+import { generateGufo } from './helpers';
+import OntoumlFactory from './ontouml_factory';
 
 describe('Enumeration', () => {
   let result;
